+++ src/main/java/ladder/view/OutputView.java
@@ -5,7 +5,7 @@ import static java.util.stream.Collectors.joining;
 import ladder.domain.Line;
 import ladder.domain.Player;
 import ladder.domain.Step;
-import ladder.domain.strategy.linestrategy.Result;
+import ladder.domain.Result;
 import java.util.List;
 
 public class OutputView {
