+++ src/test/java/ladder/domain/strategy/linestrategy/ResultTest.java
@@ -3,6 +3,7 @@ package ladder.domain.strategy.linestrategy;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
+import ladder.domain.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
